Address comments by @trustin / Build a dummy server when service is added
