+++ grpc/src/main/java/com/linecorp/armeria/server/grpc/ProtoReflectionServiceInterceptor.java
@@ -16,20 +16,6 @@
 
 package com.linecorp.armeria.server.grpc;
 
-import static com.google.common.collect.ImmutableMap.toImmutableMap;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Function;
-
-import javax.annotation.Nullable;
-
-import com.google.common.collect.ImmutableList;
-
-import com.linecorp.armeria.server.ServiceRequestContext;
-
 import io.grpc.Context;
 import io.grpc.Contexts;
 import io.grpc.InternalServer;
@@ -39,98 +25,20 @@ import io.grpc.ServerCall;
 import io.grpc.ServerCall.Listener;
 import io.grpc.ServerCallHandler;
 import io.grpc.ServerInterceptor;
-import io.grpc.ServerServiceDefinition;
 
 final class ProtoReflectionServiceInterceptor implements ServerInterceptor {
 
     static final ProtoReflectionServiceInterceptor INSTANCE = new ProtoReflectionServiceInterceptor();
-
-    @Nullable
-    private static Server dummyServer;
 
     private ProtoReflectionServiceInterceptor() {}
 
     @Override
     public <I, O> Listener<I> interceptCall(ServerCall<I, O> call, Metadata headers,
                                             ServerCallHandler<I, O> next) {
-        if (dummyServer == null) {
-            synchronized (INSTANCE) {
-                dummyServer = newDummyServer();
-            }
-        }
+        final Server server = FramedGrpcService.dummyServer;
+        assert server != null;
 
-        final Context context = Context.current().withValue(InternalServer.SERVER_CONTEXT_KEY, dummyServer);
+        final Context context = Context.current().withValue(InternalServer.SERVER_CONTEXT_KEY, server);
         return Contexts.interceptCall(context, call, headers, next);
-    }
-
-    private static Server newDummyServer() {
-        final Map<String, ServerServiceDefinition> grpcServices =
-                ServiceRequestContext.current().config().server().config().virtualHosts().stream()
-                                     .flatMap(host -> host.serviceConfigs().stream())
-                                     .map(serviceConfig -> serviceConfig.service().as(FramedGrpcService.class))
-                                     .filter(Objects::nonNull)
-                                     .flatMap(service -> service.services().stream())
-                                     // Armeria allows the same service to be registered multiple times at
-                                     // different paths, but proto reflection service only supports a single
-                                     // instance of each service so we dedupe here.
-                                     .collect(toImmutableMap(def -> def.getServiceDescriptor().getName(),
-                                                             Function.identity(),
-                                                             (a, b) -> a));
-
-        return new Server() {
-            @Override
-            public Server start() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public List<ServerServiceDefinition> getServices() {
-                return ImmutableList.copyOf(grpcServices.values());
-            }
-
-            @Override
-            public List<ServerServiceDefinition> getImmutableServices() {
-                // NB: This will probably go away in favor of just getServices above, so we
-                // implement both the same.
-                // https://github.com/grpc/grpc-java/issues/4600
-                return getServices();
-            }
-
-            @Override
-            public List<ServerServiceDefinition> getMutableServices() {
-                // Armeria does not have the concept of mutable services.
-                return ImmutableList.of();
-            }
-
-            @Override
-            public Server shutdown() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public Server shutdownNow() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public boolean isShutdown() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public boolean isTerminated() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public boolean awaitTermination(long timeout, TimeUnit unit) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public void awaitTermination() {
-                throw new UnsupportedOperationException();
-            }
-        };
     }
 }
